\name{aratio-package}
\alias{aratio-package}
\alias{aratio}
\docType{package}
\title{
Assessment ratio analysis
}
\description{
Tools for analyzing assessment ratios
}
\details{
\tabular{ll}{
Package: \tab aratio \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-09\cr
License: \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel McMillen \cr
Maintainer: Daniel McMillen <mcmillen@illinois.edu>
}

